Move Token to Nodes
