+++ src/compiler/Parsing/ParseState.cs
@@ -1,4 +1,5 @@
 using Noa.Compiler.Diagnostics;
+using Noa.Compiler.Nodes;
 using SuperLinq;
 
 namespace Noa.Compiler.Parsing;
